Refactored code to move all data handling into a specific aux file for SDMX-ML writers. Fixed imports.
