+++ src/pysdmx/io/xml/sdmx21/writer/generic.py
@@ -12,12 +12,14 @@ from pysdmx.io.xml.sdmx21.writer.__write_aux import (
     ABBR_MSG,
     ALL_DIM,
     __write_header,
-    check_content_dataset,
-    check_dimension_at_observation,
     create_namespaces,
-    get_codes,
     get_end_message,
     get_structure,
+)
+from pysdmx.io.xml.sdmx21.writer.__write_data_aux import (
+    check_content_dataset,
+    check_dimension_at_observation,
+    get_codes,
     writing_validation,
 )
 from pysdmx.io.xml.sdmx21.writer.config import CHUNKSIZE
@@ -182,7 +184,10 @@ def __write_data_single_dataset(
             prettyprint=prettyprint,
         )
     else:
-        series_codes, obs_codes = get_codes(dim, dataset)
+        series_codes, obs_codes = get_codes(
+            dimension_code=dim,
+            structure=dataset.structure,  # type: ignore[arg-type]
+        )
         att_codes = [att.id for att in dataset.structure.components.attributes]
         series_att_codes = [x for x in series_codes if x in att_codes]
         obs_att_codes = [x for x in obs_codes if x in att_codes]
@@ -397,7 +402,7 @@ def write(
 
     # Checking the dimension at observation mapping
     dim_mapping = check_dimension_at_observation(
-        content, dimension_at_observation
+        datasets=content, dimension_at_observation=dimension_at_observation
     )
     header.structure = dim_mapping
     # Generating the initial tag with namespaces
